requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
reportlab>=4.2,<5.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
dj-database-url>=2.2,<3.0
psycopg2-binary>=2.9,<3.0
